+++ ErsatzTV.Core/Emby/EmbyItemType.cs
@@ -4,4 +4,6 @@ public static class EmbyItemType
 {
     public static readonly string Movie = "Movie";
     public static readonly string Show = "Series";
+    public static readonly string Season = "Season";
+    public static readonly string Episode = "Episode";
 }