+++ ErsatzTV.Core/Jellyfin/JellyfinItemType.cs
@@ -4,4 +4,6 @@ public static class JellyfinItemType
 {
     public static readonly string Movie = "Movie";
     public static readonly string Show = "Series";
+    public static readonly string Season = "Season";
+    public static readonly string Episode = "Episode";
 }