+++ ErsatzTV.Infrastructure/Jellyfin/IJellyfinApi.cs
@@ -83,7 +83,7 @@ public interface IJellyfinApi
         [Query]
         int limit = 0);
 
-    [Get("/Items")]
+    [Get("/Items?sortOrder=Ascending&sortBy=SortName")]
     public Task<JellyfinLibraryItemsResponse> GetSeasonLibraryItems(
         [Header("X-Emby-Token")]
         string apiKey,
@@ -96,9 +96,13 @@ public interface IJellyfinApi
         [Query]
         string includeItemTypes = "Season",
         [Query]
-        bool recursive = true);
+        bool recursive = true,
+        [Query]
+        int startIndex = 0,
+        [Query]
+        int limit = 0);
 
-    [Get("/Items")]
+    [Get("/Items?sortOrder=Ascending&sortBy=SortName")]
     public Task<JellyfinLibraryItemsResponse> GetEpisodeLibraryItems(
         [Header("X-Emby-Token")]
         string apiKey,
@@ -111,7 +115,11 @@ public interface IJellyfinApi
         [Query]
         string includeItemTypes = "Episode",
         [Query]
-        bool recursive = true);
+        bool recursive = true,
+        [Query]
+        int startIndex = 0,
+        [Query]
+        int limit = 0);
 
     [Get("/Items")]
     public Task<JellyfinLibraryItemsResponse> GetCollectionLibraryItems(